Split GUI functions into GUI module
GUI functions are now located in the tram_GUI.py file
