+++ .gitignore
@@ -1,0 +1,1 @@
+__pycache__/tram_GUI.cpython-36.pyc
